pyproject.toml
[project]
name = "eurostat-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "polars>=1.39.2",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
    "pyjstat",
    "pandas",
    "python-dotenv>=1.2.2",
]
